==> src/4_export_results.ipynb <==
## Exporting results as (entity1, relation, entity2, confidence)

### Import required libraries

import pandas as pd
from pathlib import Path

### Load csv

edges_df = pd.read_csv('../data/relations.csv')

### Compute confidence score

triples_df = edges_df[['source', 'relation', 'target', 'confidence']]
triples_df.head

data_path = Path("../output")
data_path.mkdir(exist_ok=True)
triples_df.to_csv('../output/triples_output.csv', index=False)
